Add files via upload
UC Berkeley Data 8 and Data 100 Webscraping Project 
Project Description: A team of researchers with a National Science Foundation grant have been studying the diversity of data science students at UC Berkeley. As part of this grant, we (UE Data Team) administer surveys to students who are involved in data science. Each term we create a seed file for the survey. One part of the seed file is undergraduate students who are teaching for Data 8 and Data 100. We obtain the names and email addresses of these undergraduates from the Data 8 and Data 100 websites.